MCP/Custom_Client/requirements.txt
langchain_mcp_adapters
langchain_google_genai
fastmcp